+++ components/03-modules/quantity-update/quantity-update.config.js
@@ -4,19 +4,19 @@ module.exports = {
       id: 'qty',
       min: '1',
       defaultValue: '1',
-      ariaLabel: 'quantity update input'
+      ariaLabel: 'Change the quantity'
     },
     minusQtyButton: {
       tag: 'button',
       class: 'button--icon quantity-update__button quantity-update__button--minus quantity-update__button--disabled',
-      attributes: 'type="button" aria-label="quantity minus button"',
+      attributes: 'type="button" aria-label="Decrease the quantity"',
       iconClass: 'button__icon quantity-update__icon',
       iconId: 'minus'
     },
     plusQtyButton: {
       tag: 'button',
       class: 'button--icon quantity-update__button quantity-update__button--plus',
-      attributes: 'type="button" aria-label="quantity plus button"',
+      attributes: 'type="button" aria-label="Increase the quantity"',
       iconClass: 'button__icon quantity-update__icon',
       iconId: 'plus'
     },