+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update label hidden style
 - Refactored `file-upload` component - removed dependecy on button styles, JS code made more roboust
 - Update accessibility features in `file-upload` component
+- Aria-label values
 
 ### Removed
 - Old catalog product modules and catalog views